requirements.txt
airsim==1.8.1
filelock==3.18.0
fsspec==2025.3.2
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack-python==0.5.6
msgpack-rpc-python==0.4.1
networkx==3.2.1
numpy==2.0.2
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
PyYAML==6.0.2
sympy==1.13.1
torch==2.6.0
tornado==4.5.3
typing_extensions==4.13.2
